little fix

+++ lab1/Lab1.h
@@ -2,5 +2,5 @@
 
 FILE* open_file(char* filename, int bin);
 int count_lines(FILE* file);
-unsigned long int count_bytes(FILE* file);
+int count_bytes(FILE* file);
 int count_words(FILE* file);
